requirements.txt
aiohttp==3.8.4
aiohttp-client-cache==0.6.1
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.26.3
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
frozenlist==1.3.3
homeassistant-api==4.1.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
msrest==0.7.1
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
Pillow==9.4.0
playsound==1.3.0
pycairo==1.23.0
pydantic==1.9.0
PyGObject==3.42.2
python-forge==18.6.0
requests==2.28.2
requests-cache==0.9.8
requests-oauthlib==1.3.1
simplejson==3.18.3
six==1.16.0
typing_extensions==4.5.0
url-normalize==1.4.3
urllib3==1.26.14
yarl==1.8.2
